(#797) Reinstate job list methods
The methods required for the job list table were deprecated and removed while sorting out the use of Date objects, but they weren't replaced with anything. For now the methods have been restored, but remain Deprecated. Maybe they can stay as they are, or maybe they need to be altered - we need to discuss it.