(#888) Make  more robust
This used to only check records with GOOD or QUESTIONABLE flags, but
if there were no such records the map would crash because it returned
NaNs. Now it falls back to checking all records. It's not possible
to have a dataset with zero records (the processing will fail before
this point), so this will always return a range of some kind.